% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScoreTable.R
\name{attach_scales}
\alias{attach_scales}
\title{Attach additional StandardScale to already created ScoreTable}
\usage{
attach_scales(x, scale)
}
\arguments{
\item{x}{A \emph{ScoreTable} object}

\item{scale}{a \emph{StandardScale} object or list of multiple \emph{StandardScale} objects}
}
\description{
Attach additional StandardScale to already created ScoreTable
}
\examples{
# having a ScoreTable with one StandardScale attached
st <- ScoreTable(FrequencyTable(HEXACO_60$HEX_C), STEN)
st$scale
names(st$table)

# possibly attach more scales to ScoreTable
st <- attach_scales(st, list(STANINE, WECHSLER_IQ))
st$scale
names(st$table)
}
