% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handle_raw_scores.R
\name{CombScaleSpec}
\alias{CombScaleSpec}
\title{Combined Scale Specification}
\usage{
CombScaleSpec(name, ..., reverse = character(0))
}
\arguments{
\item{name}{Name of the combined scale or factor}

\item{...}{\emph{ScaleSpec} objects}

\item{reverse}{character vector containing names of the underlying subscales
of factors that need to be reversed}
}
\description{
Combine multiple \emph{ScaleSpec} objects into one in regards of \code{\link[=sum_items_to_scales]{sum_items_to_scales()}}
function. Useful when one scale of factor contains items of different possible
values or if there is hierarchy of scale or factors
}
\seealso{
Other item preprocessing functions: 
\code{\link{ScaleSpec}()},
\code{\link{sum_items_to_scale}()}
}
\concept{item preprocessing functions}
