% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comp_freqtable.R
\name{comp_freqtable}
\alias{comp_freqtable}
\title{R6 class containing computed freqtable}
\description{
Computed frequency table class. On basis of raw scores of some variables,
it creates a frequency table and allows to compute scores in some standard
scale. To initialize the object of this class the 'gen_freqtable' function
is recommended.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-get_status}{\code{comp_freqtable$get_status()}}
\item \href{#method-get_data_info}{\code{comp_freqtable$get_data_info()}}
\item \href{#method-new}{\code{comp_freqtable$new()}}
\item \href{#method-get_freqtables}{\code{comp_freqtable$get_freqtables()}}
\item \href{#method-compute_scores}{\code{comp_freqtable$compute_scores()}}
\item \href{#method-get_scoretables}{\code{comp_freqtable$get_scoretables()}}
\item \href{#method-get_computed_scores}{\code{comp_freqtable$get_computed_scores()}}
\item \href{#method-get_computed_scores_ext}{\code{comp_freqtable$get_computed_scores_ext()}}
\item \href{#method-clone}{\code{comp_freqtable$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_status"></a>}}
\if{latex}{\out{\hypertarget{method-get_status}{}}}
\subsection{Method \code{get_status()}}{
Current information about object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{comp_freqtable$get_status()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
List of metrics
\itemize{
\item n: Number of observations for frequency table generation
\item kept data: Is the data kept in the object?
\item frequency tables: Are the frequency table complete or incomplete?
\item standardized scores: Are the standardized scores computed - and if yes, then: for which scales?
}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_data_info"></a>}}
\if{latex}{\out{\hypertarget{method-get_data_info}{}}}
\subsection{Method \code{get_data_info()}}{
Information about kept data
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{comp_freqtable$get_data_info()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Initialization of the 'comp_freqtable' class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{comp_freqtable$new(data, vars, id, keep_data = T)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{data to use for computing.}

\item{\code{vars}}{character vector containing column names of numeric variables that you want to calculate frequency tables for.}

\item{\code{id}}{character string containing column name for identification of the observations. Needed only if 'keep_data' is TRUE}

\item{\code{keep_data}}{do you want to keep the table for getting scores for individual observations? Defaults to TRUE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_freqtables"></a>}}
\if{latex}{\out{\hypertarget{method-get_freqtables}{}}}
\subsection{Method \code{get_freqtables()}}{
Get generated frequency tables.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{comp_freqtable$get_freqtables()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Two lists, containing actual cumulated frequency tables and their statuses
\itemize{
\item freqtables: Cumulated frequency tables for each variable specified during initialization
\item statuses: Statuses for each frequency table. 'complete' if there was at least one observation for each raw score between actual minimum and maximum. 'incomplete' if there was at least one missing raw score and they were filled automatically
}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-compute_scores"></a>}}
\if{latex}{\out{\hypertarget{method-compute_scores}{}}}
\subsection{Method \code{compute_scores()}}{
Compute scores in scale of your choice. After computation, you can get them with 'get_scoretables' method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{comp_freqtable$compute_scores(scale)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{scale}}{Either a character string indicating one of built-in scale, or list specifying the parameters of scale of your choice
\itemize{
\item built-in scales: 'sten', 'stanine', 'tanine', 'tetronic', 'weschler-iq'
\item custom scales: specify parameters within list: 'M', 'SD', 'min', 'max' and 'name' of the scale for storage within object
}}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_scoretables"></a>}}
\if{latex}{\out{\hypertarget{method-get_scoretables}{}}}
\subsection{Method \code{get_scoretables()}}{
Get computed scoring tables for all variables
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{comp_freqtable$get_scoretables(scale)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{scale}}{character string indicating for which computed scale return values}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Two lists, containing the actual tables and parameters used for calculation
\itemize{
\item tables: tables with raw score and standardized score pairs
\item params: values of 'M', 'SD', 'min' and 'max' used for calculations
}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_computed_scores"></a>}}
\if{latex}{\out{\hypertarget{method-get_computed_scores}{}}}
\subsection{Method \code{get_computed_scores()}}{
Get computed scores for specified ids or all sourced observations. You can only use this method if you chose to keep source data.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{comp_freqtable$get_computed_scores(scale, vars = NULL, ids = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{scale}}{character string defining the computed scale. Mandatory argument.}

\item{\code{vars}}{character vector declaring for which variables of the embedded data to get the standardized score. If not specified, returns values for all}

\item{\code{ids}}{for the observations in kept data. If none is provided, then it is returned for all observations.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_computed_scores_ext"></a>}}
\if{latex}{\out{\hypertarget{method-get_computed_scores_ext}{}}}
\subsection{Method \code{get_computed_scores_ext()}}{
Get computed scores for external observations
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{comp_freqtable$get_computed_scores_ext(data, scale, vars)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{data.frame containing raw scores to get standardized scores for}

\item{\code{scale}}{character string defining the computed scale}

\item{\code{vars}}{character vector declaring variables for which to get standardized score. If the variables in external data are named differently, the vector should be named accordingly.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{comp_freqtable$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
