% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompFreqtable.R
\docType{class}
\name{CompFreqtable}
\alias{CompFreqtable}
\title{R6 class for producing re-computable FreqTable}
\description{
Computed frequency table class. It can compute and store multiple
\code{\link{FrequencyTable}s} and \code{\link{ScoreTable}s} for multiple
variables containing raw score results.

After computation, it could be also used to compute new standardized scores
for provided raw scores and integrate them into stored tables.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-attach_scale}{\code{CompFreqtable$attach_scale()}}
\item \href{#method-clone}{\code{CompFreqtable$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-attach_scale"></a>}}
\if{latex}{\out{\hypertarget{method-attach_scale}{}}}
\subsection{Method \code{attach_scale()}}{
Attach new scale to the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CompFreqtable$attach_scale(scale, overwrite = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{scale}}{\code{StandardScale} object defining a scale}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CompFreqtable$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
