% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompFreqtable.R
\docType{class}
\name{CompScoreTable}
\alias{CompScoreTable}
\title{R6 class for producing easily re-computable ScoreTable}
\description{
Computed ScoreTable class. It can compute and store \code{\link{ScoreTable}s}
for multiple variables containing raw score results.

After computation, it could be also used to compute new standardized scores
for provided raw scores and integrate them into stored tables.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-attach_scale}{\code{CompScoreTable$attach_scale()}}
\item \href{#method-attach_FrequencyTable}{\code{CompScoreTable$attach_FrequencyTable()}}
\item \href{#method-clone}{\code{CompScoreTable$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-attach_scale"></a>}}
\if{latex}{\out{\hypertarget{method-attach_scale}{}}}
\subsection{Method \code{attach_scale()}}{
Attach new scale to the object. If there are any ScoreTables
already computed, score for newly-attached scale will be computed automatically.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CompScoreTable$attach_scale(scale, overwrite = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{scale}}{\code{StandardScale} object defining a scale}

\item{\code{overwrite}}{boolean indicating if the definition for a scale
of the same name should be overwritten}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-attach_FrequencyTable"></a>}}
\if{latex}{\out{\hypertarget{method-attach_FrequencyTable}{}}}
\subsection{Method \code{attach_FrequencyTable()}}{
Attach previously generated \code{FrequencyTable} for a given
variable. \code{ScoreTable} containing every attached scale will be calulcated
automatically based on every new \code{FrequencyTable}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CompScoreTable$attach_FrequencyTable(var, ft, if_exists = "append")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{var}}{String with the name of the variable}

\item{\code{ft}}{FrequencyTable to be attached}

\item{\code{if_exists}}{Action that should be taken if \code{FrequencyTable} for
given variable already exists in the object.
\itemize{
\item \code{append} recalculates existing table
\item \code{replace} replaces existing table
}}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CompScoreTable$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
