% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScoringTable.R
\name{cond_to_df}
\alias{cond_to_df}
\title{Conditions to data.frame}
\usage{
cond_to_df(cond)
}
\arguments{
\item{cond}{list of \emph{GroupConditions}}
}
\description{
Conditions to data.frame
}
\keyword{internal}
