% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StandardScale.R
\name{default_scales}
\alias{default_scales}
\alias{STEN}
\alias{STANINE}
\alias{TANINE}
\alias{TETRONIC}
\alias{WECHSLER_IQ}
\title{Default Standard Scales}
\description{
Few \code{StandardScale} objects pre-defined for usage. To create
any other, use \code{\link[=StandardScale]{StandardScale()}} function.
\itemize{
\item \strong{STEN}: M: 5.5, SD: 2, min: 1, max: 10
\item \strong{STANINE}: M: 5, SD: 2, min: 1, max: 9
\item \strong{TANINE}: M: 50, SD: 10, min: 1, max: 100
\item \strong{TETRONIC}: M: 10, SD: 4, min: 0, max: 20
\item \strong{WECHSLER_IQ}: M: 100, SD: 15, min: 40, max: 160
}
}
