% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computations.R
\name{.calc_freq_Z_table}
\alias{.calc_freq_Z_table}
\title{helper function to calculate the regular freq table}
\usage{
.calc_freq_Z_table(data, var)
}
\arguments{
\item{data}{data.frame containing source data for calculations}

\item{var}{character string with column name of the variable}
}
\value{
data.frame containing raw scores and their quantile and Z equivalent
}
\description{
helper function to calculate the regular freq table
}
