% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computations.R
\name{.get_comp_score}
\alias{.get_comp_score}
\title{helper function to get computated scores}
\usage{
.get_comp_score(raw_score, comp_table)
}
\arguments{
\item{raw_score}{raw score to get data for}

\item{comp_table}{table with computated score}
}
\description{
helper function to get computated scores
}
