% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale.R
\name{FreqtableScale}
\alias{FreqtableScale}
\title{FreqtableScale}
\usage{
FreqtableScale(name, M, SD, min, max)
}
\arguments{
\item{name}{Name of the scale}

\item{M}{Mean of the scale}

\item{SD}{Standard deviation of the scale}

\item{min}{Minimal value the scale takes}

\item{max}{Maximal value the scale takes}
}
\description{
An object with specification for scale to be used by \code{CompFreqtable} object
}
