% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comp_freqtable.R
\name{gen_freqtable}
\alias{gen_freqtable}
\title{Compute frequency table for standardized scores computing}
\usage{
gen_freqtable(data, vars, id = NULL, keep_data = T)
}
\arguments{
\item{data}{data for computing}

\item{vars}{character vector containing column names of numeric variables that you want to calculate frequency tables for}

\item{id}{character string containing column name for identification of the observations. Needed only if 'keep_data' is TRUE}

\item{keep_data}{do you want to keep the table for getting scores for individual observations? Defaults to TRUE}
}
\description{
Compute frequency table for standardized scores computing
}
