% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{handle_retain}
\alias{handle_retain}
\title{Handle retain columns}
\usage{
handle_retain(data, output, retain)
}
\arguments{
\item{data}{input data}

\item{output}{output data.frame}

\item{retain}{retain value}
}
\description{
Internal function that handles \emph{retain} columns in non-base
normalize_scores functions
}
\keyword{internal}
