% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{IPIP_NEO_300}
\alias{IPIP_NEO_300}
\title{Sample data of IPIP-NEO-300 questionnaire results}
\format{
A data frame with 13198 rows and 7 variables
\describe{
\item{sex}{sex of the participant ('M'ale or 'F'emale)}
\item{age}{age of the participant (10--98)}
\item{N}{Raw score for Neuroticism scale (63--292)}
\item{E}{Raw score for Extraversion scale (80--296)}
\item{O}{Raw score for Openness to Experience (76--298)}
\item{A}{Raw score for Agreeableness (66--292)}
\item{C}{Raw score for Consciousness (81--299)}
}
}
\usage{
IPIP_NEO_300
}
\description{
Dataset containing sample of 13198 results of IPIP-NEO-300 results from
Johnson J.A. study published at 2014, preprocessed using \code{\link[=sum_items_to_scale]{sum_items_to_scale()}}
function. It contains many observations of different ages and sexes, also
including \code{NA} values, whenever at least one of the underlying item scores were
missing.
}
\references{
Johnson, J. A. (2014). Measuring thirty facets of the five factor model with a
120-item public domain inventory: Development of the IPIP-NEO-120.
Journal of Research in Personality, 51, 78-89.
}
\keyword{datasets}
