% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_checkers.R
\name{is_stenR_classes}
\alias{is_stenR_classes}
\alias{is.GroupConditions}
\alias{is.GroupAssignment}
\alias{is.intersected}
\alias{is.ScaleSpec}
\alias{is.CombScaleSpec}
\alias{is.FrequencyTable}
\alias{is.GroupedFrequencyTable}
\alias{is.ScoreTable}
\alias{is.GroupedScoreTable}
\alias{is.ScoringTable}
\alias{is.StandardScale}
\alias{is.Simulated}
\title{Checkers for stenR S3 and R6 classes}
\usage{
is.GroupConditions(x)

is.GroupAssignment(x)

is.intersected(x)

is.ScaleSpec(x)

is.CombScaleSpec(x)

is.FrequencyTable(x)

is.GroupedFrequencyTable(x)

is.Simulated(x)

is.ScoreTable(x)

is.GroupedScoreTable(x)

is.ScoringTable(x)

is.StandardScale(x)
}
\arguments{
\item{x}{any \R object}
}
\description{
Various functions to check if given \R object is of given class. Additionally:
\itemize{
\item \code{is.intersected()} checks if the \code{GroupAssignment} object have been created
with \code{\link[=intersect_GroupAssignment]{intersect_GroupAssignment()}} and \code{GroupedFrequencyTable},
\code{GroupedScoreTable} or \code{ScoringTable} have been created with two \code{GroupConditions}
objects.
\item \code{is.Simulated()} checks if the \code{FrequencyTable} or \code{ScoreTable} have been
created on basis of simulated distribution (based on \code{\link[=SimFrequencyTable]{SimFrequencyTable()}})
}
}
