% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_checkers.R
\name{is_stenR_methods}
\alias{is_stenR_methods}
\alias{is.ScoringTable}
\title{Checkers for stenR S3 classes}
\usage{
is.ScoringTable(x)
}
\arguments{
\item{x}{any \R object}
}
\description{
Various functions to check if given \R object is of given class
}
