% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mockNAtable}
\alias{mockNAtable}
\title{Mockup NA table}
\usage{
mockNAtable(table)
}
\arguments{
\item{table}{Table to mockup}
}
\description{
Creates mockup table if none is available for given group. Used in
\code{\link[=normalize_scores_grouped]{normalize_scores_grouped()}}. All results passed through it will have their values
"normalized" into \code{NA}
}
\keyword{internal}
