% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{normalize_score_grouped}
\alias{normalize_score_grouped}
\title{Normalize scores using Grouped table}
\usage{
normalize_score_grouped(data, vars, ..., what, retain = FALSE, .dots = list())
}
\arguments{
\item{data}{\emph{data.frame} object containing raw scores}

\item{vars}{\emph{character vector} with names of columns to normalize. Length of vars
need to be the same as number of tables provided to either \code{...} or \code{.dots}}

\item{...}{\emph{GroupedFrequencyTable} or \emph{GroupedScoreTable} objects to be used
for normalization}

\item{what}{the values to get. One of either:
\itemize{
\item \code{quan} - the quantile of x in the raw score distribution
\item \code{Z} - normalized Z score for the x raw score
\item name of the scale calculated in \emph{GroupedScoreTables} provided to \code{...} or
\code{.dots} argument
}}

\item{retain}{either \emph{boolean}: \code{TRUE} if all columns in the \code{data} are to be
retained, \code{FALSE} if none, or \emph{character vector} with names of columns to be retained}

\item{.dots}{\emph{GroupedFrequencyTable} or \emph{GroupedScoreTable} objects provided
as a list, instead of individually in \code{...}.}
}
\value{
\emph{data.frame} with normalized scores
}
\description{
Normalize scores using either \emph{GroupedFrequencyTable} or
\emph{GroupedScoreTable} for one or more variables. Given \emph{data.frame} should also
contain columns used in \emph{GroupingConditions} attached to the table
}
\seealso{
Other score-normalization functions: 
\code{\link{normalize_scores_df}()},
\code{\link{normalize_score}()}
}
\concept{score-normalization functions}
