% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{qualify_to_groups}
\alias{qualify_to_groups}
\title{Qualify observations to groups for normalization}
\usage{
qualify_to_groups(data, conditions)
}
\arguments{
\item{data}{Data with observations to group by}

\item{conditions}{GroupConditions passed to normalization function}
}
\description{
Function used internally in \code{\link[=normalize_scores_grouped]{normalize_scores_grouped()}} and
\code{\link[=normalize_scores_scoring]{normalize_scores_scoring()}}
}
\keyword{internal}
