% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SLCS}
\alias{SLCS}
\title{Sample data of SLCS questionnaire results}
\format{
A data frame with 103 rows and 19 variables
\describe{
\item{user_id}{identity anonimized with 'ids::adjective_animal'}
\item{sex}{sex of the participant ('M'ale, 'F'emale or 'O'ther)}
\item{age}{age of the participant (15--68)}
\item{SLCS_1, SLCS_2, SLCS_3, SLCS_4, SLCS_5, SLCS_6, SLCS_7, SLCS_8, SLCS_9, SLCS_10,
SLCS_11, SLCS_12, SLCS_13, SLCS_14, SLCS_15, SLCS_16}{Score for each of measure items. (1--5)}
}
}
\usage{
SLCS
}
\description{
Dataset containing individual items answers of SLCS questionnaire. They
were obtained during 2020 study on Polish incidental sample.
}
\details{
All SLCS item responses can take integer values 1-5. The measure consists
of two sub-scales: Self-Liking and Self-Competence, and the General Score
can also be calculated. Below are the item numbers that are used for each
sub-scale (\code{R} near the number means that the item need to be reversed.)
\itemize{
\item Self-Liking: 1R, 3, 5, 6R, 7R, 9, 11, 15R
\item Self-Competence: 2, 4, 8R, 10R, 12, 13R, 14, 16
\item General Score: All of the above items (they need to be reversed as in sub-scales)
}
}
\keyword{datasets}
