% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale.R
\name{StandardScale}
\alias{StandardScale}
\alias{print.StandardScale}
\alias{plot.StandardScale}
\title{An object with specification for some standard scale}
\usage{
StandardScale(name, M, SD, min, max)

\method{print}{StandardScale}(object)

\method{plot}{StandardScale}(scale, n = 1000)
}
\arguments{
\item{name}{Name of the scale}

\item{M}{Mean of the scale}

\item{SD}{Standard deviation of the scale}

\item{min}{Minimal value the scale takes}

\item{max}{Maximal value the scale takes}

\item{object}{an object for which a summary is desired.}

\item{scale}{StandardScale object}

\item{n}{Number of points the plot generates. The higher the number, the more
detailed are the plots. Default to 1000 for nicely detailed plot.}
}
\value{
StandardScale object
}
\description{
An object with specification for some standard scale
}
\details{
\code{StandardScale} objects are used with \code{\link{ScoreTable}}
objects to recalculate \code{\link{FrequencyTable}} into some standardized
scale score.

There are some StandardScale defaults available out of the box. Check out the
\code{\link{default_scales}} help page for more information.

Plot method requires \code{ggplot2} package to be installed.
}
