% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScoreTable.R
\name{strip_ScoreTable}
\alias{strip_ScoreTable}
\title{Revert the ScoreTable back to FrequencyTable object.}
\usage{
strip_ScoreTable(x)
}
\arguments{
\item{x}{a \emph{ScoreTable} object}
}
\description{
Revert the ScoreTable back to FrequencyTable object.
}
\examples{
# having a ScoreTable object
st <- ScoreTable(FrequencyTable(HEXACO_60$HEX_X), TANINE)
class(st)

# revert it back to the FrequencyTable
ft <- strip_ScoreTable(st)
class(ft)
}
