% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale.R
\name{summary.FreqtableScale}
\alias{summary.FreqtableScale}
\title{Summary method for FreqtableScale object}
\usage{
\method{summary}{FreqtableScale}(object, ...)
}
\arguments{
\item{object}{an object for which a summary is desired.}

\item{...}{additional arguments affecting the summary produced.}
}
\description{
Summary method for FreqtableScale object
}
