% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScoringTable.R
\name{verify_GC_for_ST}
\alias{verify_GC_for_ST}
\title{Internal function to verify the provided conditions with conditions available
in imported \emph{ScoringTable}}
\usage{
verify_GC_for_ST(st_df, gc_df)
}
\arguments{
\item{st_df}{data.frame form of ScoreTable}

\item{gc_df}{data.frame form of GroupConditions}
}
\value{
list of \emph{GroupConditions} objects or \code{NULL}
}
\description{
Internal function to verify the provided conditions with conditions available
in imported \emph{ScoringTable}
}
\keyword{internal}
