% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScoreTable.R
\name{attach_scales}
\alias{attach_scales}
\title{Attach additional \code{StandardScale} to already created \code{ScoreTable}}
\usage{
attach_scales(st, scale)
}
\arguments{
\item{st}{A \code{ScoreTable} object}

\item{scale}{a \code{StandardScale} object or list of multiple \code{StandardScale} objects}
}
\description{
Attach additional \code{StandardScale} to already created \code{ScoreTable}
}
\examples{
# having a ScoreTable with one StandardScale attached
st <- ScoreTable(FrequencyTable(HEXACO_60$HEX_C), STEN)
st$scale
names(st$table)

# possibly attach more scales to ScoreTable
st <- attach_scales(st, list(STANINE, WECHSLER_IQ))
st$scale
names(st$table)
}
