% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScoringTable.R
\name{check_score_between}
\alias{check_score_between}
\title{Internal function for picking up the individual value from ScoringTable}
\usage{
check_score_between(x, col_raw, col_score)
}
\arguments{
\item{x}{raw score}

\item{col_raw}{Column of raw scores from ScoringTable}

\item{col_score}{COlumn of StandardScale score from ScoringTable}
}
\description{
Internal function for picking up the individual value from ScoringTable
}
\keyword{internal}
