% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScoringTable.R
\name{export_ScoringTable}
\alias{export_ScoringTable}
\title{Export ScoringTable}
\usage{
export_ScoringTable(
  table,
  out_file,
  method = c("csv", "json"),
  conditions_file
)
}
\arguments{
\item{table}{A \emph{ScoringTable} object to export}

\item{out_file}{Output file}

\item{method}{Method for export, either \code{csv} or \code{json}}

\item{conditions_file}{Output file for \emph{GroupConditions}. Necessary only
if \code{method = csv}.}
}
\description{
After creation of \emph{ScoringTable} it can be handy to export it
into universally recognized and readable format. Two formats are currently
supported: \code{csv} and \code{json}. They can be imported back into \emph{ScoringTable}
using \code{import_ScoreTable} function.
\itemize{
\item \code{csv} format is more universally readable - it can be opened, edited
and altered (eg. before publication) in any spreadsheet manager. In case of
\emph{ScoringTable} created from \emph{GroupedScoreTable}, \emph{GroupConditions} can
be exported to another \code{csv} file, creating two different files.
\item \code{json} format can be more obtuse, but it allows export of both
\emph{ScoringTable} itself and \emph{GroupConditions} in the same \code{json} file.
}
}
