% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freqtable.R
\name{FrequencyTable}
\alias{FrequencyTable}
\alias{print.FrequencyTable}
\alias{plot.FrequencyTable}
\alias{summary.FrequencyTable}
\title{Create a FrequencyTable from raw data}
\usage{
FrequencyTable(data)

\method{print}{FrequencyTable}(ft, max = NULL)

\method{plot}{FrequencyTable}(ft)

\method{summary}{FrequencyTable}(ft)
}
\arguments{
\item{data}{vector of raw scores. Double values are coerced to integer}

\item{ft}{A \code{FrequencyTable} object}

\item{max}{numeric or NULL, specifying the maximal number of entries to be
printed. By default, when NULL, \code{\link{getOption}("max.print")} used.}
}
\value{
FrequencyTable object. Consists of:
\itemize{
\item table: data.frame with number of observations (\code{n}), frequency in sample
(\code{freq}), quantile (\code{quan}) and normalized Z-score (\code{Z}) for each point in
raw score
}
}
\description{
Create a FrequencyTable from raw data
}
