% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FrequencyTable.R
\name{GroupedFrequencyTable}
\alias{GroupedFrequencyTable}
\alias{print.GroupedFrequencyTable}
\alias{summary.GroupedFrequencyTable}
\title{Create GroupedFrequencyTable}
\usage{
GroupedFrequencyTable(data, conditions, var, force_disjoint = FALSE)

\method{print}{GroupedFrequencyTable}(gft)

\method{summary}{GroupedFrequencyTable}(gft)
}
\arguments{
\item{data}{source \emph{data.frame}}

\item{conditions}{up to two \emph{GroupConditions} objects. These objects will be
passed along during creation of higher-level objects and used when
\code{\link[=normalize_score_grouped]{normalize_score_grouped()}} will be called. If two objects are provided,
then intersection of groups will be made.}

\item{var}{name of variable to compute \emph{GroupedFrequencyTable} for}

\item{force_disjoint}{\emph{boolean}. It is recommended to keep it as default
\code{FALSE}, unless the sample size is very big and it is completely mandatory
to have the groups disjointed.}

\item{gft}{A \code{GroupedFrequencyTable} object}
}
\description{
Using \code{\link[=GroupConditions]{GroupConditions()}} object and source \emph{data.frame} compute
a set of \code{\link[=FrequencyTable]{FrequencyTable()}}s for single variable
}
\details{
\code{force_exhaustive} will always be checked as \code{FALSE} during the
calculations. It is mandatory for validity of the created \emph{FrequencyTables}
}
\seealso{
plot.GroupedFrequencyTable
}
