% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScoreTable.R
\name{GroupedScoreTable}
\alias{GroupedScoreTable}
\title{Create GroupedScoreTable}
\usage{
GroupedScoreTable(table, scale)
}
\arguments{
\item{table}{\emph{GroupedFrequencyTable} object}

\item{scale}{a \code{StandardScale} object or list of multiple \code{StandardScale} objects}
}
\value{
\emph{GroupedScoreTable} object, which constist of named \emph{list} of
\emph{ScoreTable} objects and \emph{GroupConditions} object used for grouping
}
\description{
Create GroupedScoreTable
}
\seealso{
plot.GroupedScoreTable
}
