% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScoringTable.R
\name{import_ScoringTable}
\alias{import_ScoringTable}
\title{Import ScoringTable}
\usage{
import_ScoringTable(
  source_file,
  method = c("csv", "json"),
  cond_file,
  conditions
)
}
\arguments{
\item{source_file}{Path to the file to import the \emph{ScoringTable} from}

\item{method}{Method for import, either \code{csv} or \code{json}}

\item{cond_file}{File to import the \emph{GroupConditions} from, if using \code{csv} method}

\item{conditions}{If no conditions file is provided (when using \code{csv} method)
or the existing conditions (when using \code{json} method) are to be overwritten,
provide \emph{GroupCondition} object or list of up to two of them here.}
}
\value{
ScoringTable object
}
\description{
ScoringTable can be imported from \code{csv} or \code{json} file
into R object. Source file can be either an output of \code{\link[=export_ScoreTable]{export_ScoreTable()}}
function, or created by hand - though it needs to be created following the
correct format.
}
