% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoretable.R
\name{ScoreTable}
\alias{ScoreTable}
\alias{print.ScoreTable}
\title{Create a ScoreTable using \code{FrequencyTable} and \code{StandardScale} objects}
\usage{
ScoreTable(ft, scale)

\method{print}{ScoreTable}(st, max = NULL)
}
\arguments{
\item{ft}{\code{FrequencyTable} object}

\item{scale}{\code{StandardScale} object}

\item{st}{A \code{ScoreTable} object}

\item{max}{numeric or NULL, specifying the maximal number of entries to be
printed. By default, when NULL, \code{\link{getOption}("max.print")} used.}
}
\description{
Creates a table to recalculate raw scores into specified
standardized scale. Uses normalization provided by \code{FrequencyTable}.

\code{plot.ScoreTable} method requires \code{ggplot2} package to be installed.
}
