% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comp_freqtable.R
\name{summary.CompFreqtable}
\alias{summary.CompFreqtable}
\title{Summary method for CompFreqtable object}
\usage{
\method{summary}{CompFreqtable}(object)
}
\description{
Summary method for CompFreqtable object
}
